start write encoder source

+++ src/device/NtDeviceInterface.h
@@ -8,7 +8,7 @@ class NtDeviceInterface
 {
 public:
     virtual size_t read(char *buffer, size_t bufferSize) = 0;
-    virtual int getFd() = 0;
+    virtual long getFd() = 0;
     virtual unsigned long getBufferSize() = 0;
     virtual int getWidth() { return -1; }
     virtual int getHeight() { return -1; }
